% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{aparmd-package}
\alias{aparmd-package}
\title{R Markdown templates for APA documents}
\description{
R Markdown templates for APA documents
}
\author{
Jason Bryer \email{jason@bryer.org}
}
\keyword{Rmarkdown}
\keyword{apa}
\keyword{knitr}
\keyword{markdown}

